requirements.txt
accelerate
anthropic
bitsandbytes
fireworks_ai
google-generativeai
huggingface_hub
standard-imghdr
inflect
litellm
networkx
numpy
openai
pandas
Pillow
protobuf
python-dotenv
PyYAML
scikit_learn
scipy
torch
tqdm
transformers
vllm
